Fixed login, deleted extra imports, and deleted extra getters/setters
