+++ app/src/main/AndroidManifest.xml
@@ -45,9 +45,6 @@
             android:exported="false" />
 
         <activity
-            android:name=".StoreActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true" >
             <intent-filter>
