+++ app/src/main/java/cr/ac/ucr/ecci/arceshopping/model/Products.java
@@ -6,41 +6,10 @@ public class Products {
     public ArrayList<Product> products;
     public int total;
     public int skip;
-
-
     public int limit;
-
 
     public ArrayList<Product> getProducts() {
         return products;
-    }
-
-    public void setProducts(ArrayList<Product> products) {
-        this.products = products;
-    }
-
-    public int getTotal() {
-        return total;
-    }
-
-    public void setTotal(int total) {
-        this.total = total;
-    }
-
-    public int getSkip() {
-        return skip;
-    }
-
-    public void setSkip(int skip) {
-        this.skip = skip;
-    }
-
-    public int getLimit() {
-        return limit;
-    }
-
-    public void setLimit(int limit) {
-        this.limit = limit;
     }
 
     @Override
